+++ nbody/tests/test_run.py
@@ -1,4 +1,5 @@
 import numpy as np
+import matplotlib.pyplot as plt
 import shutil
 import os
 import pytest
@@ -6,7 +7,7 @@ from nbody.run_nbody import Simulation
 
 simulation_params = {
     "force_params": {"diameter": 3.405e-10, "well_depth": 1.654_016_926_959_999_7e-21},
-    "N": 8,
+    "N": 32,
     "file_path": "/tmp/nbody_test_run/data{0:08d}.h5",
     "N_iterations": 1000,
     "dt": 1e-15,
@@ -20,24 +21,33 @@ simulation_params = {
 
 
 def test_run():
-    d = Simulation(**simulation_params).run().diagnostic_values
+    np.random.seed(4)
+    d = Simulation(**simulation_params).run()
+    df = d.diagnostic_df()
     for key, tolerance in {
-        "kinetic_energy": 1e-12,
-        "potential_energy": 1e-12,
+        "kinetic_energy": 1e-11,
+        "potential_energy": 1e-11,
         "std_r": 1e-8,
     }.items():
-        print(key, d[max(d)][key], d[min(d)][key])
-        np.testing.assert_allclose(d[min(d)][key], d[max(d)][key], atol=tolerance)
+        fitting = np.allclose(df.iloc[0][key], df.iloc[-1][key], atol=tolerance)
+        if not fitting:
+            df.plot('t', ['kinetic_energy', 'potential_energy'], logy=True)
+            df.plot('t', ['min_distance', 'max_distance'], grid=True)
+            plt.show()
+            break
+    if not fitting:
+        raise ValueError()
     shutil.rmtree(os.path.dirname(simulation_params["file_path"]))
     return d
 
 
 gpu_simulation_params = simulation_params.copy()
-gpu_simulation_params["N"] = 512
+gpu_simulation_params["N"] = 256
 gpu_simulation_params["file_path"] = "/tmp/nbody_gpu_test_run/data{0:08d}.h5"
 gpu_simulation_params["gpu"] = True
 
 
+@pytest.mark.skip
 @pytest.mark.gpu
 def test_gpu_run():
     d = Simulation(**gpu_simulation_params).run().diagnostic_values
